src/users/requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==5.0.6
django-allauth==0.63.3
djangorestframework==3.15.2
filelock==3.15.4
flake8==7.1.0
identify==2.5.36
mccabe==0.7.0
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.7.1
psycopg==3.1.19
pycodestyle==2.12.0
pyflakes==3.2.0
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
virtualenv==20.26.3
